requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.48
pydantic-settings==2.0.0
psycopg2-binary==2.9.7
pg8000==1.30.1

alembic revision --autogenerate -m "create tables"
alembic upgrade head